% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell_workflow.R
\name{clusters_seurat}
\alias{clusters_seurat}
\title{Seurat clusters from PCA embedding}
\usage{
clusters_seurat(pca_embedding, resolution)
}
\arguments{
\item{pca_embedding}{PARAM_DESCRIPTION}

\item{resolution}{PARAM_DESCRIPTION}
}
\value{
Factor with cluster identities of each cell.
}
\description{
For small data sets, quickly get Seurat clusters.
}
\details{
Type \code{clusters_seurat} and hit enter to see the code, that's what
the function is intended for - as a cheat sheet to copy paste.
If you want, you can use it to actually compute clusters, but note that
the dense distance matrix is computed using \code{dist}, so is very slow
and memory-intensive.
Requires Seurat3 to be installed.
}
\seealso{
\code{\link[Seurat]{FindNeighbors}}
}
