% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{legend_off}
\alias{legend_off}
\title{Hide Legend in ggplot2 graphs}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 1.
}
\usage{
legend_off
}
\description{
Add to plot with \code{p+legend_off}, this will set legend.position="none".
}
\keyword{datasets}
