% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{legend_point_size}
\alias{legend_point_size}
\title{Set legend size for geom_point}
\usage{
legend_point_size(size = 2)
}
\arguments{
\item{size}{Point size in legend, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
\code{geom_point} with small size makes the legend unintelligible,
add to plot with \code{p+legend_point_size(3)}.
}
\details{
Same as ggplot2::guides.
}
\examples{
ggplot(mtcars, aes(mpg, cyl, col=cyl))+geom_point() + legend_point_size(5)
}
\seealso{
\code{\link[ggplot2]{c("guide_bins", "guide_colourbar", "guide_coloursteps", "guide_legend", "guides", "guides")}},\code{\link[ggplot2]{guide_legend}}
}
