% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\docType{data}
\name{plain_umap}
\alias{plain_umap}
\title{Plain UMAP theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 5.
}
\usage{
plain_umap
}
\description{
A ggplot theme that can be added with \code{p+plain_umap}.
Simple background, no axis ticks etc..
}
\keyword{datasets}
